Move CreateSymbolicLink
